Relicense project from MIT to GPLv2
The rust library doomgeneric includes [doomgeneric](https://github.com/ozkl/doomgeneric)
in it's binary (and this projects includes that library in turn).
Due to the nature of GPLv2, it seems to be illegal to have
a MIT licensed project include GPLv2 licensed code directly.

This should solve the issue.
I'm the only contributer so far, so there is no one
else to ask confirmation for this change from.